Add windows build
This adds the typical `build`/`test`/`upload` steps for our new windows
buildkite workers.  Our windows buildkite workers have `bash` and `make`
and all the typical goodies pre-installed within docker containers that
are managed in the rootfs-images repository [0], which allows us to keep
the toolchain versions independent of the base windows image.

I have copied over many of the data files used to create the Julia
installer, so that when we run codesigning and such, we aren't making
use of any code paths within the main Julia repository.

[0] https://github.com/JuliaCI/rootfs-images/tree/main/windows