Disable PowerPC CI
And though we reached the summit, we did not discover salvation, only misery....

Summit is being turned off tomorrow. As far as we know, that was the last PowerPC platform that anybody was seriously using Julia on. CI has been broken on master recently and nobody seems particularly inclined to fix it, so turn off CI and we'll drop down PowerPC to Tier 4.