Mark mmtk CI tests as softfail
As I said in https://github.com/JuliaCI/julia-buildkite/pull/419#issuecomment-2635689784,
mmtk should be softfail unless and until the root scanner is moved into the main
julia repo so that ABI changes don't depend on an external repo. Otherwise we
would slow down base Julia evolution for an experimental feature, which is not
reasonable.